Delete Type Conversion

+++ src/test/anovos/drift/test_stability.py
@@ -97,7 +97,7 @@ def test_that_stability_index_can_be_calculated(
 
     df_stability.index = df_stability["attribute"]
     assert_almost_equal(
-        df_stability.loc["A", cols_to_check_numerical].astype("float"),
+        df_stability.loc["A", cols_to_check_numerical],
         [0.132, 0.507, 0.162, 2.0, 0.0, 2.0, 1.4, 0.0],
         3,
     )
